Update README.md
In this project, the default password "4567" is defined in the Arduino, but it can be changed through coding. Upon entering a password, the system verifies it with the stored one. If correct, it displays 'Access Granted, Welcome,' and the servo motor rotates the door to 180 degrees, unlocking it. Five seconds are provided to enter through the door, after which it automatically locks.

In case of a wrong password, the display shows 'Code Incorrect, Go Away,' and the buzzer beeps once with each key press.